Fixing log. Inverting !cmd and !!cmd behavior. Removing char setting from .conf.
Besides being verbose fprintf(stderr, buf) is wrong because it doesn't escape %. fprintf(stderr, "%s", buf) should be used if there was no vfprintf.

Invert !cmd and !!cmd so that !!cmd suppress the confirmation. Suppressing the confirmation can be seen as the less common use. The confirmation is potentially inconvenient if there are multiple bots in the channel or if there is a desire to conceal the fact that a user is a bot. These cases can be seen as less common.

Removing char setting from .conf because the setting is rarely desired. Alternatively the .conf char should be changed from > to ! because that's the prefix in the readme examples and the main.cc default prefix.

Fixing indentation in readme.